+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Released on Sunday, August 23 2026
 
 - Fixed matching of `annotation-xml` encoding to be case-insensitive (#1284) @arpitjain099
+- Fixed `<xmp>` usage in `<select>` elements
 
 # 1.7.1